refactor: refactor Stack e2e
